+++ packages/spacecat-shared-data-access/src/dto/configuration.js
@@ -26,6 +26,7 @@ export const ConfigurationDto = {
       version: dynamoItem.version,
       queues: dynamoItem.queues,
       jobs: dynamoItem.jobs,
+      handlers: dynamoItem.handlers,
     };
 
     return createConfiguration(configurationData);
@@ -41,5 +42,6 @@ export const ConfigurationDto = {
     version: configuration.getVersion(),
     queues: configuration.getQueues(),
     jobs: configuration.getJobs(),
+    handlers: configuration.getHandlers(),
   }),
 };