+++ packages/spacecat-shared-data-access/src/index.d.ts
@@ -84,49 +84,12 @@ export interface Audit {
   getScores: () => object;
 }
 
-/**
- * AuditConfigType defines the structure for specific audit type configurations.
- */
-export interface AuditConfigType {
-  /**
-   * Returns true if the audit type is disabled for the site. If an audit type is disabled, no
-   * audits of that type will be scheduled for the site.
-   * @returns {boolean} True if the audit type is disabled for the site.
-   */
-  disabled: () => boolean;
-}
-
 export interface Config {
 
 }
 
 export interface FulfillableItems {
   items: string[];
-}
-
-/**
- * AuditConfig defines the structure for the overall audit configuration of a site.
- */
-export interface AuditConfig {
-  /**
-   * Returns true if audits are disabled for the site. If audits are disabled, no audits will be
-   * scheduled for the site. Overrides any audit type specific configurations.
-   * @returns {boolean} True if audits are disabled for the site.
-   */
-  auditsDisabled: () => boolean;
-
-  /**
-   * Returns the audit config for a specific audit type. The audit type is the key.
-   * @param {string} auditType The audit type to get the config for.
-   * @returns {AuditConfigType} The audit config for the audit type.
-   */
-  getAuditTypeConfig: (auditType: string) => AuditConfigType;
-
-  /**
-   * Returns the audit configs for all audit types. The keys are the audit types.
-   * @returns {object} The audit configs for all audit types.
-   */
-  getAuditTypeConfigs: () => object;
 }
 
 /**
@@ -221,12 +184,6 @@ export interface Site {
    * @returns {string} The last update timestamp.
    */
   getUpdatedAt: () => string;
-
-  /**
-   * Retrieves the current audit configuration for the site.
-   * @returns {AuditConfig} The current audit configuration.
-   */
-  getAuditConfig: () => AuditConfig;
 
   /**
    * Retrieves the current configuration for the site.
@@ -434,6 +391,60 @@ export interface Configuration {
    * @returns {Array} The jobs configurations.
    */
   getJobs: () => Array<object>;
+
+    /**
+     * Retrieves the handlers configuration.
+     * @returns {object} The handlers configuration.
+     */
+  getHandlers: () => object;
+
+  /**
+   * Retrieves the handler configuration for handler type.
+   * @returns {object} The handler type configuration.
+   */
+  getHandler: (type) => object;
+
+  /**
+   * Return true if a handler type is enabled for an organization.
+   * @param type handler type
+   * @param org organization
+   */
+  isHandlerEnabledForOrg: (type: string, org: Organization) => boolean;
+
+  /**
+   * Return true if a handler type is enabled for a site.
+   * @param type handler type
+   * @param site site
+   */
+  isHandlerEnabledForSite: (type: string, site: Site) => boolean;
+
+  /**
+   * Enables a handler type for an site.
+   * @param type handler type
+   * @param site site
+   */
+  enableHandlerForSite: (type: string, site: Site) => void;
+
+  /**
+   * Enables a handler type for an organization.
+   * @param type handler type
+   * @param org organization
+   */
+  enableHandlerForOrg: (type: string, org: Organization) => void;
+
+    /**
+     * Disables a handler type for an site.
+     * @param type handler type
+     * @param site site
+     */
+  disableHandlerForSite: (type: string, site: Site) => void;
+
+  /**
+   * Disables a handler type for an organization.
+   * @param type handler type
+   * @param org organization
+   */
+  disableHandlerForOrg: (type:string, org: Organization) => void;
 
 }
 
@@ -682,6 +693,10 @@ export interface DataAccess {
   removeOrganization: (
       organizationId: string,
   ) => Promise<void>;
+  getImportJobsByDateRange: (
+      startDate: string,
+      endDate: string,
+  ) => Promise<ImportJob[]>;
   getImportJobByID: (
     id: string,
     ) => Promise<ImportJob | null>;
@@ -757,6 +772,7 @@ interface DataAccessConfig {
   indexNameAllOrganizations: string,
   indexNameAllOrganizationsByImsOrgId: string,
   indexNameAllImportJobsByStatus: string,
+  indexNameAllImportJobsByDateRange: string,
   indexNameAllImportUrlsByJobIdAndStatus: string,
   pkAllSites: string;
   pkAllLatestAudits: string;