+++ packages/spacecat-shared-rum-api-client/test/functions.test.js
@@ -15,12 +15,14 @@ import { expect } from 'chai';
 import cwv from '../src/functions/cwv.js';
 import notfound from '../src/functions/404.js';
 import experiment from '../src/functions/experiment.js';
+import trafficAcquisition from '../src/functions/traffic-acquisition.js';
 import variant from '../src/functions/variant.js';
 import bundles from './fixtures/bundles.json' assert { type: 'json' };
 import bundlesForVariant from './fixtures/bundles_for_variant.json' assert { type: 'json' };
 import expectedCwvResult from './fixtures/cwv.json' assert { type: 'json' };
 import expected404Result from './fixtures/notfound.json' assert { type: 'json' };
 import expectedExperimentsResult from './fixtures/experiments.json' assert { type: 'json' };
+import expectedTrafficSourcesResult from './fixtures/trafficSources.json' assert { type: 'json' };
 import expectedVariantResult from './fixtures/variant.json' assert { type: 'json' };
 
 describe('Query functions', () => {
@@ -42,5 +44,10 @@ describe('Query functions', () => {
   it('crunches variant data', async () => {
     const variantResult = variant.handler(bundlesForVariant.rumBundles);
     expect(expectedVariantResult).to.eql(variantResult);
+  });
+
+  it('crunches traffic acquisition', async () => {
+    const trafficSourcesResult = trafficAcquisition.handler(bundles.rumBundles);
+    expect(expectedTrafficSourcesResult).to.eql(trafficSourcesResult);
   });
 });