+++ packages/spacecat-shared-dynamo/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@adobe/spacecat-shared-dynamo-v1.3.29](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.28...@adobe/spacecat-shared-dynamo-v1.3.29) (2024-07-22)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#295](https://github.com/adobe/spacecat-shared/issues/295)) ([4e8d50d](https://github.com/adobe/spacecat-shared/commit/4e8d50d8d88b2d68b483fda2ead31542d15ee952))
+
+# [@adobe/spacecat-shared-dynamo-v1.3.28](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.27...@adobe/spacecat-shared-dynamo-v1.3.28) (2024-07-13)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#291](https://github.com/adobe/spacecat-shared/issues/291)) ([5678328](https://github.com/adobe/spacecat-shared/commit/5678328a8efe4c2068b3b796274d8e4392f75c8a))
+
 # [@adobe/spacecat-shared-dynamo-v1.3.27](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-dynamo-v1.3.26...@adobe/spacecat-shared-dynamo-v1.3.27) (2024-07-08)